requirements/prod.txt
# Flask and plugins
flask==0.12.2
nexmo==2.3.0
M2Crypto==0.32.0
validatesns==0.1.1
python-dotenv==0.10.1